Segregate regression and non-regression tests
